04 was updated by Aleksandrov

+++ labs/504/aleksandrov.yuriy/04_virtual_memory/memory.cpp
@@ -102,13 +102,13 @@ int main()
         for (int i = 0; i < LDT_size; ++i)
             cin >> hex >> LDT[i];
         
-        uint16_t /* ???? */ page_directory_size;
+        uint32_t page_directory_size;
         cin >> dec >> page_directory_size;
         vector<uint64_t> page_directory(page_directory_size);
         for (int i = 0; i < page_directory_size; ++i)
             cin >> hex >> page_directory[i];
         
-        uint16_t /* ???? */ page_table_size;
+        uint32_t page_table_size;
         cin >> dec >> page_table_size;
         vector<uint64_t> page_table(page_table_size);
         for (int i = 0; i < page_table_size; ++i)
